funcionalidades funcionando  correctamente

+++ src/config/socket.js
@@ -6,6 +6,8 @@ const ProductRepository = require('../dao/productRepository.js');
 const messageManager = new MessageRepository();
 const productManager = new ProductRepository();
 const CustomError = require('../services/errors/CustomError.js');
+const { sendProductDeletionEmail } = require('../public/js/emailProductEliminado.js');
+const User = require('../dao/models/user.model.js');
 
 // Configura el middleware de sesión
 const sessionMiddleware = session({
@@ -30,7 +32,7 @@ const initializeSocket = (httpServer) => {
             socket.emit('error', { status: 'error', message: 'Usuario no autenticado' });
             return;
         }
-        console.log('Usuario autenticado:', user); // Verificar si la información del usuario está disponible
+
         // Manejo de mensajes
         messageManager.readMessage()
             .then((messages) => {
@@ -50,8 +52,10 @@ const initializeSocket = (httpServer) => {
                             socket.emit('responseAdd', 'Mensaje enviado');
                         });
                 })
-                .catch((error) =>
-                    socket.emit('responseAdd', 'Error al enviar el mensaje: ' + error.message));
+                .catch((error) => {
+                    console.error("Error al enviar el mensaje:", error.message);
+                    socket.emit('responseAdd', 'Error al enviar el mensaje: ' + error.message);
+                });
         });
 
         // Manejo de productos
@@ -59,6 +63,7 @@ const initializeSocket = (httpServer) => {
             .then((products) => {
                 socket.emit('products', products);
             });
+
         socket.on('NewProduct', (product) => {
             if (!user) {
                 socket.emit('error', { status: 'error', message: 'Usuario no autenticado' });
@@ -76,7 +81,7 @@ const initializeSocket = (httpServer) => {
                         });
                 })
                 .catch((error) => {
-                    console.log("error al agregar el producto", error)
+                    console.error("Error al agregar el producto:", error.message);
                     if (error instanceof CustomError) {
                         socket.emit('error', { status: 'error', error: error.name, message: error.message, cause: error.cause });
                     } else {
@@ -85,30 +90,54 @@ const initializeSocket = (httpServer) => {
                 });
         });
 
-        socket.on('eliminarProduct', (product) => {
-            const userId = socket.request.session.user._id || socket.request.session.user.id; // Obtener el ID del usuario
-            const userRole = socket.request.session.user.role; // Obtener el rol del usuario
-            if (!userId) {
-                socket.emit('error', { status: 'error', message: 'Usuario no autenticado' });
-                return;
-            }
-        if (!product._id) {
-            throw new Error('ID del producto no proporcionado');
-        }
+        socket.on('eliminarProduct', async (product) => {
+            try {
+                const userId = socket.request.session.user._id || socket.request.session.user.id;
+                const userRole = socket.request.session.user.role;
 
-            productManager.deleteProduct(product._id, userId, userRole)
-                .then(() => {
-                    return productManager.readProducts();
-                })
-                .then((products) => {
-                    socket.emit('products', products);
-                    socket.emit('responseDelete', 'Producto eliminado');
-                })
-                .catch((error) => {
-                    socket.emit('responseDelete', 'Error al eliminar el producto: ' + error.message);
-                });
+                if (!userId) {
+                    socket.emit('error', { status: 'error', message: 'Usuario no autenticado' });
+                    return;
+                }
+
+                if (!product._id) {
+                    throw new Error('ID del producto no proporcionado');
+                }
+
+                // Verificar si el producto existe y si el usuario tiene permiso para eliminarlo
+                const productToDelete = await productManager.getProductById(product._id);
+                if (!productToDelete) {
+                    throw new Error('El producto no pudo ser encontrado');
+                }
+
+                if (userRole !== 'admin' && productToDelete.owner.toString() !== userId) {
+                    throw new Error('No tienes permiso para eliminar este producto');
+                }
+
+                // Eliminar el producto
+                const deleteResult = await productManager.deleteProduct(product._id, userId, userRole);
+                if (deleteResult.deletedCount === 0) {
+                    throw new Error('El producto no pudo ser eliminado');
+                }
+
+                // Obtener el dueño del producto antes de eliminar
+                const owner = await User.findById(productToDelete.owner);
+                if (owner && owner.role === 'premium') {
+                    const ownerEmail = owner.email;
+                    const ownerName = owner.nombre;
+                    await sendProductDeletionEmail(ownerEmail, ownerName, product._id);
+                }
+
+                // Actualizar la lista de productos
+                const products = await productManager.readProducts();
+                socket.emit('products', products);
+                socket.emit('responseDelete', 'Producto eliminado y notificado por correo.');
+            } catch (error) {
+                console.error('Error al eliminar el producto:', error.message);
+                socket.emit('responseDelete', 'Error al eliminar el producto: ' + error.message);
+            }
         });
-        
+
         socket.on('editarProduct', (data) => {
             const { id, updatedProduct } = data;
             if (!user) {
@@ -116,7 +145,17 @@ const initializeSocket = (httpServer) => {
                 return;
             }
 
-            productManager.updateProduct(id, updatedProduct)
+            // Verificar que el usuario sea dueño del producto o admin
+            productManager.getProductById(id)
+                .then((existingProduct) => {
+                    if (!existingProduct) {
+                        throw new Error('Producto no encontrado');
+                    }
+                    if (user.role !== 'admin' && existingProduct.owner.toString() !== user._id) {
+                        throw new Error('No tienes permiso para editar este producto');
+                    }
+                    return productManager.updateProduct(id, updatedProduct);
+                })
                 .then(() => {
                     productManager.readProducts()
                         .then((products) => {
@@ -124,8 +163,10 @@ const initializeSocket = (httpServer) => {
                             socket.emit('responseEdit', 'Producto actualizado');
                         });
                 })
-                .catch((error) =>
-                    socket.emit('responseEdit', 'Error al actualizar el producto: ' + error.message));
+                .catch((error) => {
+                    console.error('Error al actualizar el producto:', error.message);
+                    socket.emit('responseEdit', 'Error al actualizar el producto: ' + error.message);
+                });
         });
     });
